+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Development
 
 * Better error message for bayesian/bootstrap models when `hypotheses()` attempts to post-process a previous `marginaleffects` call.
+* Model matrix is attached to more `newdata`. Useful for Issue #6 in `marginaleffectsJAX`.
 
 ## 0.28.0